backend/src/helper/Twilio.py
from twilio.rest import Client
import os
from dotenv import load_dotenv

load_dotenv()

account_sid = os.getenv('TWILIO_ACCOUNT_SID')
auth_token = os.getenv('TWILIO_AUTH_TOKEN')
twilio_phone_number = os.getenv('TWILIO_PHONE_NUMBER')

client = Client(account_sid, auth_token)


def SendMessage(body: str, to: str):
    client.messages.create(
        from_=twilio_phone_number,
        body=body,
        to=to
    )
